[#56] count 타입 변경(Integer > Long, Float > Double), 상세 서비스 구현 진행 중

+++ src/main/java/com/depromeet/breadmapbackend/bakeries/controller/BakeriesController.java
@@ -62,7 +62,7 @@ public class BakeriesController {
     @PostMapping(value = "/{bakeryId}/rating")
     public ResponseEntity<Void> registerBakeryRating(@PathVariable Long bakeryId, @RequestBody RegisterBakeryRatingRequest registerBakeryRatingRequest){
         RegisterBakeryRatingResponse registerBakeryRatingResponse = new RegisterBakeryRatingResponse();
-        float totalRating = registerBakeryRatingResponse.getRating();
+        Double totalRating = registerBakeryRatingResponse.getRating();
         totalRating = totalRating/registerBakeryRatingRequest.getRating();
         return ApiResponse.created(null);
     }
